requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy
pandas
matplotlib
scikit-learn
joblib
flask
flask-restful
fastapi
uvicorn[standard]
nltk
unicodedata
seaborn